G1.5: split resolver/infer.rs — pure line-scanners → infer_lines.rs
Move all index-free functions to src/resolver/infer_lines.rs:
- extract_collection_element_type, first_type_arg
- infer_type_in_lines, infer_type_in_lines_raw
- find_rhs_str, infer_from_rhs_assignment
- has_dot_after_first_call
- extract_return_type_from_detail, extract_type_with_generics
- find_declaration_range_in_lines

resolver/infer.rs retains only index-backed functions (ReceiverKind,
ReceiverType, infer_receiver_type, infer_variable_type*, infer_field_type*,
find_field_type_in_class, infer_method_return_type, find_*_return_type).

Update call sites:
- lines_ext.rs: infer_type/infer_type_raw/find_declaration_range → infer_lines::
- resolver/mod.rs: re-export extract_collection_element_type from infer_lines
- infer_tests.rs: import from super::super::infer_lines directly

830 tests pass, 0 new clippy warnings.
