+++ src/resolver/mod.rs
@@ -8,6 +8,7 @@ pub(crate) mod find;
 mod hierarchy;
 mod import_edit;
 pub(crate) mod infer;
+pub(crate) mod infer_lines;
 pub(crate) mod resolve;
 #[cfg(test)]
 mod tests;
@@ -20,10 +21,8 @@ pub(crate) use complete::{
 };
 pub(crate) use hierarchy::walk_hierarchy;
 pub(crate) use import_edit::{already_imported, import_insertion_line, make_import_edit};
-pub(crate) use infer::{
-    extract_collection_element_type, infer_receiver_type, infer_variable_type_raw, ReceiverKind,
-    ReceiverType,
-};
+pub(crate) use infer::{infer_receiver_type, infer_variable_type_raw, ReceiverKind, ReceiverType};
+pub(crate) use infer_lines::extract_collection_element_type;
 pub(crate) use resolve::{
     ensure_file_data, fqns_for_name, resolve_symbol_inner, resolve_symbol_no_rg,
 };
@@ -40,9 +39,10 @@ use fd::import_file_stems;
 #[cfg(test)]
 use fd::package_prefix;
 #[cfg(test)]
-pub(crate) use infer::{
+pub(crate) use infer::infer_variable_type;
+#[cfg(test)]
+pub(crate) use infer_lines::{
     find_declaration_range_in_lines, infer_type_in_lines, infer_type_in_lines_raw,
-    infer_variable_type,
 };
 #[cfg(test)]
 pub(crate) use resolve::{is_stdlib, resolve_symbol};
